feat(install.sh): Allow specifying install directory.  Do not ask for sudo password if not required.

+++ install.sh
@@ -10,6 +10,8 @@ error() {
     exit 1
 }
 
+INSTALL_DIR="/usr/local/bin"
+
 OS=$(uname -s)
 
 ARCH=$(uname -m)
@@ -24,8 +26,59 @@ case $ARCH in
     i386) ARCH="i386";;
 esac
 
+# Print usage instructions.
+printUsage() {
+  cat <<END
+usage: $(basename "$0") [-h] [--dir DIR]
+
+Installs "${EXECUTABLE}" from https://github.com/${GITHUB_USER}/${GITHUB_PROJECT}.
+
+Optional Arguments:
+    -h        - Show this help.
+    --dir DIR - Select the directory to install the executable to.
+
+END
+}
+
+while :; do
+    case $1 in
+        -h|-\?|--help)   # Call a "show_help" function to display a synopsis, then exit.
+            printUsage
+            exit
+            ;;
+        -d|--dir)       # Takes an option argument, ensuring it has been specified.
+            if [ -n "$2" ]; then
+                INSTALL_DIR=$2
+                shift
+            else
+                printf 'ERROR: "--dir" requires a non-empty option argument.\n' >&2
+                exit 1
+            fi
+            ;;
+        -?*)
+            printf 'Unknown option: "%s"\n\n' "$1" >&2
+            printUsage >&2
+            exit 1
+            ;;
+        *)               # Default case: If no more options then break out of the loop.
+            break
+    esac
+
+    shift
+done
+
+if [ $# -ne 0 ]; then
+    printUsage >&2
+    exit 1
+fi
+
 if ! command -v curl > /dev/null; then
     error "curl must be installed to run this script."
+fi
+
+if [ ! -d "${INSTALL_DIR}" ]; then
+    printf 'Installation directory %s does not exist' "${INSTALL_DIR}" >&2
+    exit 1
 fi
 
 # Download the archive and unpack it to /tmp.
@@ -37,13 +90,16 @@ if [ ! -e "/tmp/${EXECUTABLE}" ]; then
     error "Download failed."
 fi
 
-echo "Copying ${EXECUTABLE} to /usr/local/bin."
-sudo -p "Enter your password to install ${EXECUTABLE}:" mv "/tmp/${EXECUTABLE}" /usr/local/bin || {
-    rm "/tmp/${EXECUTABLE}"
-    error "Failed to copy ${EXECUTABLE} to /usr/local/bin." >&2
-}
+DEST="${INSTALL_DIR}/${EXECUTABLE}"
+echo "Copying ${EXECUTABLE} to '${INSTALL_DIR}'."
+if ! mv "/tmp/${EXECUTABLE}" "${DEST}" >  /dev/null 2>&1 ; then
+    sudo -p "Enter your password to install ${EXECUTABLE}:" mv "/tmp/${EXECUTABLE}" "${DEST}" || {
+        rm "/tmp/${EXECUTABLE}"
+        error "Failed to copy ${EXECUTABLE} to '${INSTALL_DIR}'." >&2
+    }
+fi
 
-echo "Installed ${EXECUTABLE} to /usr/local/bin"
+echo "Installed ${EXECUTABLE} to '${INSTALL_DIR}'."
 
 # Show setup instructions
-/usr/local/bin/${EXECUTABLE} setup
+"${INSTALL_DIR}/${EXECUTABLE}" setup
